Remove Overview/CRD and Overview/ResourceUtil extensions
