+++ frontend/packages/dev-console/src/plugin.tsx
@@ -20,8 +20,6 @@ import {
   Perspective,
   RoutePage,
   OverviewResourceTab,
-  OverviewCRD,
-  OverviewResourceUtil,
   YAMLTemplate,
   OverviewTabSection,
   ReduxReducer,
@@ -43,10 +41,6 @@ import {
 } from '@console/internal/models';
 import * as models from './models';
 import { getKebabActionsForKind } from './utils/kebab-actions';
-import {
-  getPipelinesAndPipelineRunsForResource,
-  tknPipelineAndPipelineRunsResources,
-} from './utils/pipeline-plugin-utils';
 import { FLAG_OPENSHIFT_PIPELINE, ALLOW_SERVICE_BINDING, FLAG_OPENSHIFT_GITOPS } from './const';
 import {
   newPipelineTemplate,
@@ -61,7 +55,6 @@ import * as yamlIcon from './images/yaml.svg';
 import * as importGitIcon from './images/from-git.svg';
 import * as dockerfileIcon from './images/dockerfile.svg';
 import * as pipelineIcon from './images/pipeline.svg';
-import { operatorResources } from './components/topology/operators/operator-resources';
 import {
   HelmTopologyConsumedExtensions,
   helmTopologyPlugin,
@@ -101,8 +94,6 @@ type ConsumedExtensions =
   | ReduxReducer
   | KebabActions
   | OverviewResourceTab
-  | OverviewCRD
-  | OverviewResourceUtil
   | YAMLTemplate
   | OverviewTabSection
   | AddAction
@@ -267,24 +258,6 @@ const plugin: Plugin<ConsumedExtensions> = [
     },
     flags: {
       required: [FLAGS.OPENSHIFT],
-    },
-  },
-  {
-    type: 'Overview/CRD',
-    properties: {
-      resources: tknPipelineAndPipelineRunsResources,
-    },
-    flags: {
-      required: [FLAG_OPENSHIFT_PIPELINE],
-    },
-  },
-  {
-    type: 'Overview/ResourceUtil',
-    properties: {
-      getResources: getPipelinesAndPipelineRunsForResource,
-    },
-    flags: {
-      required: [FLAG_OPENSHIFT_PIPELINE],
     },
   },
   {
@@ -1157,15 +1130,6 @@ const plugin: Plugin<ConsumedExtensions> = [
           verb: 'create',
         },
       ],
-    },
-  },
-  {
-    type: 'Overview/CRD',
-    properties: {
-      resources: operatorResources,
-    },
-    flags: {
-      required: [ALLOW_SERVICE_BINDING],
     },
   },
   {
